feat: Add functionality to Upload products and pictures via Csv File

+++ OgmentoAPI.Domain.Common.Infrastructure/Repository/PictureRepository.cs
@@ -4,6 +4,7 @@ using OgmentoAPI.Domain.Common.Abstractions.CustomExceptions;
 using OgmentoAPI.Domain.Common.Abstractions.DataContext;
 using OgmentoAPI.Domain.Common.Abstractions.Models;
 using OgmentoAPI.Domain.Common.Abstractions.Repository;
+using System.Security.Cryptography;
 
 
 namespace OgmentoAPI.Domain.Common.Infrastructure.Repository
@@ -28,7 +29,7 @@ namespace OgmentoAPI.Domain.Common.Infrastructure.Repository
 			List<Picture> pictures = _dbContext.Picture.Where(x=>pictureIds.Contains(x.PictureID)).ToList();
 			return pictures.Select(x => new PictureModel
 			{
-				Id = x.PictureID,
+				PictureId = x.PictureID,
 				FileName = x.FileName,
 				MimeType = x.MimeType,
 				Hash = _dbContext.PictureBinary.Single(b => b.PictureId==x.PictureID).Hash,
@@ -52,6 +53,11 @@ namespace OgmentoAPI.Domain.Common.Infrastructure.Repository
 				PictureId = pictureEntity.Entity.PictureID,
 				BinaryData =pictureModel.BinaryData,
 			};
+			using (MD5 md5 = MD5.Create())
+			{
+				byte[] hashBytes = md5.ComputeHash(pictureModel.BinaryData);
+				pictureBinary.Hash = BitConverter.ToString(hashBytes).Replace("-", "").ToLowerInvariant();
+			}
 			EntityEntry<PictureBinary> pictureBinaryEntity = _dbContext.PictureBinary.Add(pictureBinary);
 			int rowsAdded = await _dbContext.SaveChangesAsync();
 			if (rowsAdded == 0)
@@ -59,7 +65,7 @@ namespace OgmentoAPI.Domain.Common.Infrastructure.Repository
 				throw new DatabaseOperationException("Unable to add pictures.");
 			}
 			pictureModel.Hash = pictureBinaryEntity.Entity.Hash;
-			pictureModel.Id = pictureEntity.Entity.PictureID;
+			pictureModel.PictureId = pictureEntity.Entity.PictureID;
 			return pictureModel;
 		}
 		public async Task DeletePicture(string? hash)
@@ -81,7 +87,12 @@ namespace OgmentoAPI.Domain.Common.Infrastructure.Repository
 
 		public async Task DeletePictures(List<int> pictureIds)
 		{
-			int rowsDeleted = await _dbContext.Picture.Where(x => pictureIds.Contains(x.PictureID)).ExecuteDeleteAsync();
+			int rowsDeleted = await _dbContext.PictureBinary.Where(x => pictureIds.Contains(x.PictureId)).ExecuteDeleteAsync();
+			if (rowsDeleted == 0)
+			{
+				throw new DatabaseOperationException("Unable to delete the Picture Binary.");
+			}
+			rowsDeleted = await _dbContext.Picture.Where(x => pictureIds.Contains(x.PictureID)).ExecuteDeleteAsync();
 			if (rowsDeleted == 0)
 			{
 				throw new DatabaseOperationException("Unable to delete the pictures.");
